+++ changelog.md
@@ -1,6 +1,7 @@
 # Next
 - [BUG] Fixed issue with paranoid deletes and `deletedAt` with a custom field.
 - [BUG] No longer crahes on `where: []`
+- [FEATURE] Validations are now enabled by default for upsert.
 
 # 2.0.0-rc7
 - [FEATURE] Throw an error if no where clause is given to `Model.destroy()`.